+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Added more cart information to the example templates to help with debugging.
 - Added client side polling to the example templates to detect if the current cart has changes on another tab or session.
+- Added `craft\commerce\services\Carts::getHasSessionCartNumber()`.
 
 ### Fixed
 - Fixed a bug where `craft/commerce/elements/Order::EVENT_BEFORE_ADD_LINE_ITEM` had the `isNew` param set incorrectly. ([#851](https://github.com/craftcms/commerce/pull/851))